alfred_json/requirements.txt
jsonschema ~= 2.6
pycontracts ~= 1.8.0
